sysinfo: fix overflow in base::SleepForNanoseconds()
Found this while playing around with block_manager-stress-test: I saw that
a 15s sleep turned into a ~2s sleep.

I switched SleepForMilliseconds() over to int64_t for good measure.
Previously it'd overflow if you tried to sleep for ~24 days.

Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5942
Tested-by: Jean-Daniel Cryans <jdcryans@cloudera.com>

+++ src/kudu/gutil/sysinfo.cc
@@ -76,7 +76,7 @@ namespace base {
 static double cpuinfo_cycles_per_second = 1.0;  // 0.0 might be dangerous
 static int cpuinfo_num_cpus = 1;  // Conservative guess
 
-void SleepForNanoseconds(int nanoseconds) {
+void SleepForNanoseconds(int64_t nanoseconds) {
   // Sleep for nanosecond duration
   struct timespec sleep_time;
   sleep_time.tv_sec = nanoseconds / 1000 / 1000 / 1000;
@@ -85,7 +85,7 @@ void SleepForNanoseconds(int nanoseconds) {
     ;  // Ignore signals and wait for the full interval to elapse.
 }
 
-void SleepForMilliseconds(int milliseconds) {
+void SleepForMilliseconds(int64_t milliseconds) {
   SleepForNanoseconds(milliseconds * 1000 * 1000);
 }
